+++ src/cfnlint/rules/jsonschema/Base.py
