+++ src/cfnlint/rules/jsonschema/__init__.py
@@ -1,1 +1,1 @@
-from cfnlint.rules.jsonschema.base import BaseJsonSchema
+from cfnlint.rules.jsonschema.Base import BaseJsonSchema
