+++ src/cfnlint/rules/resources/Configuration.py
@@ -5,7 +5,7 @@ SPDX-License-Identifier: MIT-0
 import cfnlint.helpers
 from cfnlint.data.schemas.other import resources
 from cfnlint.jsonschema import CfnTemplateValidator, ValidationError
-from cfnlint.rules.jsonschema.base import BaseJsonSchema
+from cfnlint.rules.jsonschema.Base import BaseJsonSchema
 from cfnlint.schema.manager import PROVIDER_SCHEMA_MANAGER
