+++ src/cfnlint/rules/resources/iam/Policy.py
@@ -8,7 +8,7 @@ import json
 
 from cfnlint.helpers import load_resource
 from cfnlint.jsonschema import RefResolver
-from cfnlint.rules.jsonschema.base import BaseJsonSchema
+from cfnlint.rules.jsonschema.Base import BaseJsonSchema
 
 
 # pylint: disable=unused-argument
